refactor: restructure forms controllers
